requirements.txt
#eun 
#pip install -r requirements.txt
opencv-python
pytesseract
python-dotenv
PyMuPDF
python-docx
opencv-python
pytesseract
python-dotenv
ollama
streamlit
PyMuPDF
python-docx
opencv-python
pytesseract
python-dotenv

